refactor: rename uploader implementation modules for consistency
