+++ mithril-aggregator/src/file_uploaders/mod.rs
@@ -1,13 +1,13 @@
-mod dumb_snapshot_uploader;
+mod dumb_file_uploader;
 mod file_uploader;
-mod local_snapshot_uploader;
-mod remote_snapshot_uploader;
+mod local_file_uploader;
+mod remote_file_uploader;
 
-pub use dumb_snapshot_uploader::*;
+pub use dumb_file_uploader::*;
 pub use file_uploader::FileLocation;
 pub use file_uploader::FileUploader;
-pub use local_snapshot_uploader::LocalSnapshotUploader;
-pub use remote_snapshot_uploader::RemoteSnapshotUploader;
+pub use local_file_uploader::LocalSnapshotUploader;
+pub use remote_file_uploader::RemoteSnapshotUploader;
 
 #[cfg(test)]
 pub use file_uploader::MockFileUploader;
